% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_prism.R
\name{download_prism}
\alias{download_prism}
\title{Download PRISM Climate Data maps}
\usage{
download_prism(
  sp_res = "4km",
  data = "ppt",
  start_date = as.Date("2017-04-01"),
  end_date = as.Date("2017-08-03"),
  t_res = "monthly",
  out_dir = paste0(getwd()),
  pause = 1
)
}
\arguments{
\item{sp_res}{a character vector that denotes the spatial resolution of the
climate maps wanted}

\item{data}{a character vector that is one of the following:
ppt, tmin, tmax, tmean, tdmean, vpdmin, or vpdmax}

\item{start_date}{a character vector or date object in the format YYYY-MM-DD}

\item{end_date}{a character vector or date object in the format YYYY-MM-DD}

\item{t_res}{a character vector that denotes the time resolution
of the map. (daily, monthly, or yearly)}

\item{out_dir}{out_dir is the name of a output directory or folder that will
be created in the working directory. The User can input a name for the
in which the data will be saved in.}

\item{pause}{an integer value that denotes the amount of seconds to pause
between each download}
}
\description{
Download PRISM Climate Data maps
}
\details{
This function creates a folder and downloads PRISM data
for the selected date(s), map type(s), and data map that the user provides.
The downloaded data is stored in the permanent folder.
}
