% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ghcnd.R
\name{get_country_ids}
\alias{get_country_ids}
\title{Get all station ids for a given country}
\usage{
get_country_ids(file_path, country_code)
}
\arguments{
\item{file_path}{The file path where your ghcnd-stations.txt is located}

\item{country_code}{WMO country code}
}
\value{
A character vector of station ids
}
\description{
Get all station ids for a given country
}
