% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sens_farms.R
\name{get_sens_farms}
\alias{get_sens_farms}
\title{Get the Sen's slope estimate and p-value given an sf dataframe of farm locations}
\usage{
get_sens_farms(farms_coords, sap_prop)
}
\arguments{
\item{farms_coords}{An sf dataframe containing farms and their coordinates.}

\item{sap_prop}{A terra SpatRaster containing yearly sap day proportion
rasters, probably calculated from \code{\link[=sap_day]{sap_day()}}.}
}
\value{
An sf dataframe containing the time series of sap day proportions
for each year, the geometry, the Sen's slope estimate, and the Sen's slope
p-value.
}
\description{
This function takes an sf dataframe of (assumed) farm coordinates. Then, it
extracts the time series of yearly sap day proportions for each farm. Sen's
slope is calculated for each time series. The slope estimate and p-value are
added as columns to the extracted data frame.
}
