% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sens_significance.R
\name{get_sens_significance}
\alias{get_sens_significance}
\title{Add a significance column}
\usage{
get_sens_significance(farms_sf)
}
\arguments{
\item{farms_sf}{An sf dataframe containing the time series of sap day proportions
for each year, the geometry, the Sen's slope estimate, and the Sen's slope
p-value, probably outputted by \code{\link[=get_sens_farms]{get_sens_farms()}}.}
}
\value{
The same sf dataframe passed into the function, but with the added
significance column.
}
\description{
Add a Sen's slope significance column given an sf dataframe probably outputted
by \code{\link[=get_sens_farms]{get_sens_farms()}}. Using the previously calculated Sen's slope already
present in the dataframe, a significance variable is
created that is 1 if the slope is significant-positive, -1 if the slope is
negative-significant, and 0 if the slope is not significant.
}
