% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sap_day.R
\name{sap_day}
\alias{sap_day}
\title{Calculate Ideal Sap Days from Temperature Rasters}
\usage{
sap_day(tmax_rast, tmin_rast, t_upper = 2.2, t_lower = -1.1)
}
\arguments{
\item{tmax_rast}{A raster stack of tmax values that must have terra::time
values containing date information to subset by year.}

\item{tmin_rast}{A raster stack of tmin values that must have terra::time
values containing date information to subset by year.}

\item{t_upper}{The upper temperature value for the freeze/thaw cycle that
will be used in a logical statement to find ideal sap tapping days.}

\item{t_lower}{The lower temperature value for the freeze/thaw cycle that
will be used in a logical statement to find ideal sap tapping days.}
}
\value{
A list of length two - a raster stack of the proportion of ideal
sap tapping days per year, and a raster stack of the sum of ideal sap
tapping days per year.
}
\description{
Maple sap flows the most on days where temperatures are oscillating between
freezing and thawing, known as the freeze/thaw cycle. This function uses
tmax and tmin rasters (potentially loaded in from \code{\link[=loca_t_rast]{loca_t_rast()}} or
\code{\link[=prism_t_rast]{prism_t_rast()}}) to calculate the proportion and sum of days in each year
that are ideal for maple sap tapping. Using a 30-year record of sap flow
and weather observations at the Uihlein Forest, Lake Placid, NY, sap flow
occurred on 80\% of the days when the minimum temperature fell below
-1.1 °C and maximum temperature exceeded 2.2 °C, which is where the
default values for the temperature threshold come from (Skinner et al.
2010).
}
